+++ isafe/__main__.py
@@ -1,0 +1,6 @@
+from .isafe import run
+
+if __name__ == '__main__':
+    run()
+    
+